Commit Feb 23 2023

+++ download_with_session_ID.py
@@ -1327,12 +1327,13 @@ def download_files_with_mastersessionlist(sessionlist_filename,masktype,filetype
             # print("CSV FILE URL::{}".format(each_selected_scan[masktype+'_'+filetype+'FILENAME']))
             this_filename=os.path.join(dir_to_save,os.path.basename(each_selected_scan[masktype+'_'+filetype+'FILENAME']))
             download_a_singlefile_with_URIString(each_selected_scan[masktype+'_'+filetype+'FILENAME'],os.path.basename(each_selected_scan[masktype+'_'+filetype+'FILENAME']),dir_to_save)
-            this_filename_df=pd.read_csv(this_filename)
-            this_filename_df['SESSION_ID']=each_selected_scan[masktype+'_'+filetype+'FILENAME'].split('/')[3]
+            if ".csv" in this_filename:
+                this_filename_df=pd.read_csv(this_filename)
+                this_filename_df['SESSION_ID']=each_selected_scan[masktype+'_'+filetype+'FILENAME'].split('/')[3]
             # this_filename_df['FILEPATH'+filetype]=each_selected_scan[masktype+'_'+filetype+'FILENAME'] #.split('/')[3]
             # listofsession_current_df_row=listofsession_current_df[listofsession_current_df['SESSION_ID']==each_selected_scan[masktype+'_'+filetype+'FILENAME'].split('/')[3]]
             # print("listofsession_current exists::{}".format(listofsession_current_df_row  ))
-            this_filename_df.to_csv(this_filename,index=False)
+                this_filename_df.to_csv(this_filename,index=False)
             files_local_location.append(this_filename)
         print("I SUCCEEDED AT ::{}".format(inspect.stack()[0][3]))
         return files_local_location
